fix(observability): address CI ruff failures
- Move record_lag_sample import to top of compound_finalize (PLC0415)
- Drop redundant int() around round() in schedule_lag (RUF046)
- Annotate the websocket-manager local import in system.py with a
  noqa + explanation (api→services layer inversion would require
  moving ConnectionManager to services/websocket.py, deferred).
- ruff format pass on pod_stats.py + system.py

+++ backend/app/services/schedule_lag.py
@@ -129,7 +129,7 @@ def compute_window(window_seconds: int) -> ScheduleLagStats:
 
     def percentile(p: float) -> int:
         # Nearest-rank percentile (same convention as RED's percentiles).
-        idx = max(0, min(n - 1, int(round(p * (n - 1)))))
+        idx = max(0, min(n - 1, round(p * (n - 1))))
         return lags[idx]
 
     return ScheduleLagStats(
